feat: adicionar modificações no carrino ,entitie, repositorie e service

+++ src/main/java/br/edu/ufersa/projeto9poo/models/repositories/CarrinhoRepository.java
@@ -12,7 +12,7 @@ public interface CarrinhoRepository {
 
     void deletar(Carrinho carrinho);
 
-    Optional<Carrinho> buscar(int id);
+    Optional<Carrinho> buscar(Long id);
 
     List<Carrinho> buscarTodos();
 }
